requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
pydantic
pypdf
tiktoken
openai
jinja2
python-multipart
python-dotenv
pgvector>=0.2.5
syntok==1.4.4
